+++ packages/odata-service-inquirer/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ux/odata-service-inquirer
 
+## 0.5.30
+
+### Patch Changes
+
+-   f7d2895: Adds fallback to v4 catalog service for auth or not found errors
+
+## 0.5.29
+
+### Patch Changes
+
+-   34b2606: Adds new system prompts for Abap on BTP
+
+## 0.5.28
+
+### Patch Changes
+
+-   Updated dependencies [d3dafeb]
+    -   @sap-ux/btp-utils@0.15.2
+    -   @sap-ux/axios-extension@1.16.5
+    -   @sap-ux/telemetry@0.5.24
+
 ## 0.5.27
 
 ### Patch Changes